caption parsing for koef

+++ bot.py
@@ -9,12 +9,12 @@ from stylizer import stylize
 
 API_TOKEN = '5503345880:AAF_QgSgOHwrMlEY5ABAWajpsDhXxxe3DxM'
 
-# Configure logging
 logging.basicConfig(level=logging.INFO)
 
-# Initialize bot and dispatcher
 bot = Bot(token=API_TOKEN)
 dp = Dispatcher(bot)
+
+__all__ = ['set_style', 'set_base', 'get_style', 'get_base', 'run', 'upscale', ]
 
 
 @dp.message_handler(commands='start')
@@ -28,7 +28,7 @@ async def send_welcome(message: types.Message):
 
 
 @dp.message_handler(commands='help')
-async def send_welcome(message: types.Message):
+async def help(message: types.Message):
     """
     Вывести листинг комманд бота
     """
@@ -45,14 +45,17 @@ async def download_img(path:str, message: types.Message):
     Функция скачивания картинки
     :param path: путь, куда надо сохранить картинку
     :param message: объект сообщения от пользователя
-    :return:
+    :return: размер картинки
     """
     if message.photo:
         await message.photo[-1].download(destination_file=path)
+        return message.photo[-1]['width'], message.photo[-1]['height']
     elif message.document:
         await message.document.download(destination_file=path)
+        return message.document['thumb']['width'], message.document['thumb']['height']
     else:
         await bot.send_message(message.chat.id, 'Прикрепите картинку')
+        return None
 
 
 @dp.message_handler(commands='style', commands_ignore_caption=False, content_types=["document", "photo"])
@@ -112,26 +115,37 @@ async def run(message: types.Message):
         await bot.send_message(message.chat.id, "Сначала определите исходные картинки: (/style - стиля, /base - основы)")
 
 
-@dp.message_handler(content_types=["document", "photo"])
+@dp.message_handler(content_types=["document", "photo", "text"])
 async def upscale(message: types.Message):
     """
-    Увеличить картинку в 4 раза
+    Увеличить картинку в указанное кол-во раз (по умолчанию 4)
     """
     model = arch.RRDBNet(3, 3, 64, 23, gc=32)
     model.load_state_dict(torch.load('RRDB_ESRGAN_x4.pth'), strict=True)
     model.eval()
-    await download_img(f'{message.chat.id}.png', message)
-    img = cv2.imread(f'{message.chat.id}.png', cv2.IMREAD_COLOR)
-    img = img * 1.0 / 255
-    img = torch.from_numpy(np.transpose(img[:, :, [2, 1, 0]], (2, 0, 1))).float()
-    img_LR = img.unsqueeze(0)
-    with torch.no_grad():
-        output = model(img_LR).data.squeeze().float().cpu().clamp_(0, 1).numpy()
-    output = np.transpose(output[[2, 1, 0], :, :], (1, 2, 0))
-    output = (output * 255.0).round()
-    cv2.imwrite(f'{message.chat.id}.png', output)
-    await bot.send_photo(message.chat.id, types.InputFile(f'{message.chat.id}.png'))
-    os.remove(f'{message.chat.id}.png')
+    img_size = await download_img(f'{message.chat.id}.png', message)
+    if img_size:
+        print(img_size)
+        print(message.caption)
+        if message.caption:
+            text = message.caption
+            if '-k' in text:
+                koef = text.split('-k')[1][: text.rfind(' ')].strip()
+                print(koef)
+                if not koef.isdigit():
+                    await bot.send_message(message.chat.id, "Коэффицент должен быть числом")
+                    return
+        img = cv2.imread(f'{message.chat.id}.png', cv2.IMREAD_COLOR)
+        img = img * 1.0 / 255
+        img = torch.from_numpy(np.transpose(img[:, :, [2, 1, 0]], (2, 0, 1))).float()
+        img_LR = img.unsqueeze(0)
+        with torch.no_grad():
+            output = model(img_LR).data.squeeze().float().cpu().clamp_(0, 1).numpy()
+        output = np.transpose(output[[2, 1, 0], :, :], (1, 2, 0))
+        output = (output * 255.0).round()
+        cv2.imwrite(f'{message.chat.id}.png', output)
+        await bot.send_photo(message.chat.id, types.InputFile(f'{message.chat.id}.png'))
+        os.remove(f'{message.chat.id}.png')
 
 
 if __name__ == '__main__':
